+++ python/GafferTractor/TractorDispatcher.py
@@ -146,6 +146,14 @@ class TractorDispatcher( GafferDispatch.Dispatcher ) :
 
 		task = self.__acquireTask( batch, dispatchData )
 		tractorParent.addChild( task )
+		if isinstance( tractorParent.subtasks[-1], GafferTractor.tractorAPI().Instance ) :
+			# The `author` API automatically creates `Instance` objects when the
+			# same task appears as a subtask of multiple tasks. But it uses the
+			# task `title` to reference the task, and that might not be unique.
+			# An apparently undocumented feature of Tractor Engine helps us out
+			# by looking up by `id` before `title` when resolving instances. So
+			# we update the Instance to reference by `id` instead.
+			tractorParent.subtasks[-1].title = task.id
 
 		if batch.blindData().get( "tractorDispatcher:visited" ) :
 			return
@@ -169,7 +177,14 @@ class TractorDispatcher( GafferDispatch.Dispatcher ) :
 		# Make a task.
 
 		nodeName = batch.node().relativeName( dispatchData["scriptNode"] )
-		task = GafferTractor.tractorAPI().Task( title = nodeName )
+		task = GafferTractor.tractorAPI().Task(
+			title = nodeName,
+			# We need `id` to be unique so that we can reference it
+			# from `Instance` objects. TaskBatch is RefCounted, which
+			# uses the object address for `__hash__()`, giving us the
+			# uniqueness we require.
+			id = str( batch.__hash__() )
+		)
 
 		if batch.frames() :